fix: use provider
